+++ src/test/java/study/SetTest.java
@@ -1,4 +1,4 @@
-package string;
+package study;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
